Continue with d24 on wsl2.
WIP

+++ 2015/cpp/d24.h
@@ -19,7 +19,7 @@ protected:
 	void post_benchmark() override {}
 
 private:
-	using ans_t = int;
+	using ans_t = uint64_t;
 	std::pair<ans_t, ans_t> solution(const std::string_view &sv);
 	std::unique_ptr<file_contents> input_file;
 	std::vector<std::pair<std::string_view, std::pair<ans_t, ans_t>>> input = {
